backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
bcrypt
pydantic[email]
python-multipart
